backend/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pandas
pyarrow
boto3
python-dotenv
google-genai
pydantic
